perf: change DB to use GIN indexing
I removed everything that had to do with the
keyword functionality since GIN has that.
Added logging here and there, but there'll be a
whole commit for that later on.

Lots of added stuff to DBManager, lots and lots,
I'll come back to this commit when I have time to
explain, hopefully