feat: add search cache
As advised, using the proxy pattern, a cache layer has been made.
Basically, we "intercept" the request for the SearchSelector and
before we pass it on, we check a cache for the result.

An interface has also been added, in order to show it
does its job.

I made a typo in the README and refactored search-form because
it looked like ass, but was too lazy to separate the commits.